Permissions done

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\{Post,Tag};
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -31,8 +32,12 @@ class PostController extends Controller
      */
     public function create()
     {
-		$tags = Tag::all();
-        return view("post_create", ['tags' => $tags]);
+		if(Gate::authorize('create')){
+			$tags = Tag::all();
+			return view("post_create", ['tags' => $tags]);
+		} else {
+			return "Вы не авторизованы поэтому не можете писать статьи!";
+		}
     }
 
     /**
@@ -43,20 +48,25 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-		$v = $request->validate([
-			'name' => 'required|min:5|max:100',
-			'slug' => 'required|alpha_dash|unique:posts',
-			'anonce' => 'required|max:255',
-			'content' => 'required',
-		]);
-		$v["publish"] = $request->publish;
-		$post = Post::create($v);
-		if(!empty($request->tags)){
-			foreach ($request->tags as $tag) {
-				$post->tags()->attach($tag);
+		if(Gate::authorize('create')){
+			$v = $request->validate([
+				'name' => 'required|min:5|max:100',
+				'slug' => 'required|alpha_dash|unique:posts',
+				'anonce' => 'required|max:255',
+				'content' => 'required',
+			]);
+			$v["publish"] = $request->publish;
+			$v["user_id"] = $request->user_id;
+			$post = Post::create($v);
+			if(!empty($request->tags)){
+				foreach ($request->tags as $tag) {
+					$post->tags()->attach($tag);
+				}
 			}
+			return back()->with('success','Статья успешно создана');
+		} else {
+			return "Вы не авторизованы поэтому не можете писать статьи!";
 		}
-        return back()->with('success','Статья успешно создана');
     }
 
     /**
@@ -82,8 +92,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-		$tags = Tag::all();
-		return view("post_update", ['post' => $post, 'tags' => $tags]);
+		if(Gate::authorize('edit', $post)){
+			$tags = Tag::all();
+			return view("post_update", ['post' => $post, 'tags' => $tags]);
+		} else {
+			return "У вас нет прав на редактирование статьи";
+		}
     }
 
     /**
@@ -95,21 +109,25 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $v = $request->validate([
-			'name' => 'required|min:5|max:100',
-			'slug' => 'required|alpha_dash|unique:posts,id,' . $post->id,
-			'anonce' => 'required|max:255',
-			'content' => 'required',
-		]);
-		$v["publish"] = $request->publish;
-		$post->update($v);
-		$post->tags()->detach();
-		if(!empty($request->tags)){
-			foreach ($request->tags as $tag) {
-				$post->tags()->attach($tag);
+		if(Gate::authorize('edit', $post)){
+			$v = $request->validate([
+				'name' => 'required|min:5|max:100',
+				'slug' => 'required|alpha_dash|unique:posts,id,' . $post->id,
+				'anonce' => 'required|max:255',
+				'content' => 'required',
+			]);
+			$v["publish"] = $request->publish;
+			$post->update($v);
+			$post->tags()->detach();
+			if(!empty($request->tags)){
+				foreach ($request->tags as $tag) {
+					$post->tags()->attach($tag);
+				}
 			}
+			return back()->with('success','Статья успешно обновлена');
+		} else {
+			return "У вас нет прав на редактирование статьи";
 		}
-		return back()->with('success','Статья успешно обновлена');
     }
 
     /**
@@ -120,7 +138,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
-        return back()->with('success','Статья успешно удалена');
+		if(Gate::authorize('edit', $post)){
+			$post->delete();
+			return back()->with('success','Статья успешно удалена');
+		} else {
+			return "Только владелец статьи может её удалить!";
+		}
     }
 }
